backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]
python-multipart
pydantic>=2.0
pillow
numpy
scipy
aiofiles
